Build with Grunt
I don't know if you'll like this or not but why build with Grunt?

1.  It's cross platform (works on Windows)

2.  Less global installs needed

    with make I needed make installed (I guess that's common), I also
    needed coffee install, might be the wrong version, and I needed
    uglify installed.

    With Grunt those dependencies are installed insided node_modules.

    so

        git clone git://github.com/gka/chroma.js
        npm install
        grunt

    I think you can build without installing grunt as well with

        ./node_modules/grunt/bin/grunt

Also I don't really know what you wanted this to do. If you check
the current chroma.js the license looks like

    // Generated by CoffeeScript 1.6.2
    /** echo  * @license echo  * while read i do echo  *  done echo
    */

That's fixed assuming you wanted it fixed

Take it or leave it. I just thought I'd pass it on