Returning original colors when calling colorScale.colors() with no arguments
Previously colorScale.colors() with no arguments would return the colors
that correspond to the min and max values in the domain. This changes
the behavior to return the original colors, regardless of how many
colors were used to make the scale. This symmetry will make it easier to
use a colorScale as the sole representation of the scale.

This change originated by my desire to generate a css gradient from a
colorScale. Without this change, I need to provide both the colorScale
and the colors used in the scale. With this change I can generate the
css gradient purely from the colorScale (with some other assumptions).